TRAC#7506 Adds the participants app
