+++ app/participants/templates/participant_registration.html
